pyproject.toml
[tool.poetry]
name = "data-platform"
version = "0.1.0"
description = "End-to-end data platform example"
authors = ["Example Developer <developer@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
structlog = "^24.1.0"
opentelemetry-api = "^1.23.0"
opentelemetry-sdk = "^1.23.0"
opentelemetry-instrumentation-fastapi = "^0.44b0"
httpx = "^0.27.0"
prefect = "^2.14.0"
pandas = "^2.2.0"
duckdb = "^0.10.0"
celery = "^5.3.6"
redis = "^5.0.1"
flower = "^2.0.1"
streamlit = "^1.30.0"
plotly = "^5.19.0"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.5"
mypy = "^1.8.0"
ruff = "^0.1.9"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
